1 hours stock data from March 12 until July 5, 2018
the compressed file of one hour of stock data from March 12, 2018 until July 5, 2018 is committed. This data pertains to 500 stocks belonging to the nifty 500 category on the NSE for the Indian stock market.

The stock with the ticker symbol which have an & have the same name of the corresponding file